Refactoring of the serialization of the basic .rdp file.
